\name{hesc_demo}
\alias{hesc_demo}
\docType{data}
\title{HESC demo Seurat object for SUN examples}
\description{
  Small Seurat object bundled for examples and README quick start. Use with
  \code{SUN()} after building or refreshing the SNN graph as in the README workflow.
}
\usage{hesc_demo}
\format{An object of class \code{Seurat}.}
\keyword{datasets}
